+++ backend/surveys/admin_views.py
@@ -22,472 +22,7 @@ from datetime import datetime
 from surveys.tasks import analyze_survey_responses_async
 from django.db import transaction, IntegrityError
 
-
-# Use a single logger configuration
 logger = logging.getLogger("surveys")
-
-@ensure_csrf_cookie
-def set_csrf_token(request):
-    """
-    Set CSRF token for the client.
-    
-    @params:
-        request (HttpRequest): The HTTP request object
-    
-    @returns:
-        JsonResponse: JSON response containing CSRF token
-        - success (bool): Always True
-        - csrfToken (str): The CSRF token for the client
-    """
-    return JsonResponse({'detail': 'CSRF cookie set'})
-
-def demo_survey_view(request):
-    """
-    Render the demo survey page.
-    
-    @params:
-        request (HttpRequest): The HTTP request object
-    
-    @returns:
-        HttpResponse: Rendered demo survey HTML template
-    """
-    if request.method == "GET":
-        return render(request, "demo_survey.html")
-
-@api_view(["GET"])
-def user_view(request):
-    """
-    Render the user dashboard page.
-    
-    @params:
-        request (HttpRequest): The HTTP request object
-    
-    @returns:
-        HttpResponse: Rendered user dashboard HTML template
-    """
-    if request.user.is_authenticated:
-        # Create the response first
-        response = Response(UserSerializer(request.user).data)
-        
-        # Check and set missing cookies
-        if not request.COOKIES.get('auth_token'):
-            response.set_cookie(
-                'auth_token', 
-                request.session.session_key, 
-                max_age=3600*24*7, 
-                path='/', 
-                domain=settings.COOKIE_DOMAIN,
-                secure=settings.COOKIE_SECURE,
-                httponly=False,
-                samesite='Lax'
-            )
-        
-        if not request.COOKIES.get('user_email'):
-            response.set_cookie(
-                'user_email', 
-                request.user.email, 
-                max_age=3600*24*7, 
-                path='/', 
-                domain=settings.COOKIE_DOMAIN,
-                secure=settings.COOKIE_SECURE,
-                httponly=False,
-                samesite='Lax'
-            )
-            
-        if not request.COOKIES.get('is_superuser'):
-            response.set_cookie(
-                'is_superuser', 
-                'true' if request.user.is_superuser else 'false', 
-                max_age=3600*24*7, 
-                path='/', 
-                domain=settings.COOKIE_DOMAIN,
-                secure=settings.COOKIE_SECURE,
-                httponly=False,
-                samesite='Lax'
-            )
-            
-        if not request.COOKIES.get('is_institution_admin'):
-            response.set_cookie(
-                'is_institution_admin', 
-                'true' if request.user.role == User.Role.INSTITUTION_ADMIN else 'false', 
-                max_age=3600*24*7, 
-                path='/', 
-                domain=settings.COOKIE_DOMAIN,
-                secure=settings.COOKIE_SECURE,
-                httponly=False,
-                samesite='Lax'
-            )
-        
-        return response
-    else:
-        return Response({"error": "User not authenticated"}, status=401)
-
-@api_view(["POST"])
-def survey_view(request, hash_link=None):
-    """
-    Render the survey page for a specific survey template.
-    
-    @params:
-        request (HttpRequest): The HTTP request object
-        hash_link (str): Unique hash identifier for the survey template
-    
-    @returns:
-        HttpResponse: Rendered survey HTML template with context data
-        Context includes:
-        - survey_template: The survey template object
-        - questions: List of questions for the survey
-        - hash_link: The survey hash link
-    """
-    # Handle hash link survey submission
-    if hash_link:
-        try:
-            survey_template = SurveyTemplate.objects.get(hash_link=hash_link)
-            # Get all questions for this template
-            questions = SurveyQuestion.objects.filter(survey_template=survey_template)
-            if not questions.exists():
-                return JsonResponse({"success": False, "message": "No questions found in the survey template"})
-                
-            # Check if all questions have responses
-            missing_responses = []
-            for question in questions:
-                question_id = str(question.id)
-                if question_id not in request.data:
-                    missing_responses.append(question.question_text[:30] + "...")
-            
-            if missing_responses:
-                return JsonResponse({"success": False, "message": f"Missing responses for questions: {', '.join(missing_responses)}"})
-
-            # Handle the survey submission
-            return _handle_student_responses(request, survey_template, questions, True)
-            
-        except SurveyTemplate.DoesNotExist:
-            return JsonResponse({"success": False, "message": "Survey template not found"}, status=404)
-    # Check if a valid user is submitting the response
-    if not request.user.is_authenticated:
-        return JsonResponse({"success": False, "message": "Please login to the application to submit a survey response"})
-    # Get the survey template - either from request or use a default
-    survey_template_id = request.data.get('survey_template_id')
-    if not survey_template_id:
-        # Try to get the used template first
-        if request.user.institution_details:
-            try:
-                survey_template = SurveyTemplate.objects.filter(
-                    institution=request.user.institution_details,
-                    used=True
-                ).first()
-                
-                # If no used template exists, fall back to the one with minimal ID
-                if not survey_template:
-                    survey_template = SurveyTemplate.objects.filter(
-                        institution=request.user.institution_details
-                    ).order_by('id').first()
-                    
-                    # If we found a template, mark it as used
-                    if survey_template:
-                        survey_template.used = True
-                        survey_template.save()
-                
-                if not survey_template:
-                    return JsonResponse({"success": False, "message": "No survey template found for your institution"})
-            except Exception as e:
-                return JsonResponse({"success": False, "message": f"Error finding survey template: {str(e)}"})
-        else:
-            return JsonResponse({"success": False, "message": "No institution associated with user and no survey template specified"})
-    else:
-        try:
-            survey_template = SurveyTemplate.objects.get(id=survey_template_id)
-        except Exception as e:
-            return JsonResponse({"success": False, "message": f"Invalid survey template: {str(e)}"})
-
-    # Get all questions for this template
-    questions = SurveyQuestion.objects.filter(survey_template=survey_template)
-    if not questions.exists():
-        return JsonResponse({"success": False, "message": "No questions found in the survey template"})
-    
-    # Check if all questions have responses
-    missing_responses = []
-    for question in questions:
-        question_id = str(question.id)
-        if question_id not in request.data:
-            missing_responses.append(question.question_text[:30] + "...")
-    
-    if missing_responses:
-        return JsonResponse({"success": False, "message": f"Missing responses for questions: {', '.join(missing_responses)}"})
-
-    # Important: You can't return a @api_view within another @api_view
-    return _handle_student_responses(request, survey_template, questions, False)
-
-
-def _handle_student_responses(request, survey_template, questions, hashed=False):
-    """
-    Handle student survey responses and save them to the database.
-    
-    @params:
-        request (HttpRequest): The HTTP request object containing survey responses
-        survey_template (SurveyTemplate): The survey template object
-    
-    @returns:
-        JsonResponse: JSON response indicating success or failure
-        - success (bool): True if responses saved successfully, False otherwise
-        - message (str): Success or error message
-        - response_id (int): ID of the created survey response (on success)
-    """
-    no_student_user = False
-    if hashed:
-        student_name = request.data.get('student_name')
-        school_email = request.data.get('school_email')
-        institution_email_regex = survey_template.institution.institution_regex_pattern
-        if not re.fullmatch(institution_email_regex, school_email, re.IGNORECASE):
-            return JsonResponse({
-                "success": False,
-                "message": "Please use your institution email. Normally should end with .edu"
-            })
-        student = User.objects.filter(role=User.Role.STUDENT, email=school_email).first()
-        if not student:
-            no_student_user=True
-            ano_student, created = AnonymousStudent.objects.get_or_create(
-                email=school_email,
-                defaults={'name': student_name, 'survey_template': survey_template}
-            )
-        else:
-            # Registered user exists with this email
-            if not request.user.is_authenticated:
-                # They need to log in first
-                return JsonResponse({
-                    "success": False,
-                    "requires_auth": True,
-                    "message": "This email is already registered. Please log in to submit your survey.",
-                })
-            else:
-                # They're already logged in - use their account
-                student = request.user
-                no_student_user = False
-            
-    else:
-        # Get the email and validate it ends with .edu
-        school_email = request.user.email
-
-        # Save the student response
-        student = request.user
-    
-        # Update student name if provided
-        if 'student_name' in request.data and not student.name:
-            student.name = request.data['student_name']
-            student.save()
-    
-    # Create the survey response
-    try:
-        with transaction.atomic():
-            if not no_student_user:
-                recent = SurveyResponse.objects.select_for_update().filter(
-                    student=student,
-                    survey_template=survey_template,
-                    created__gte=timezone.now() - timezone.timedelta(seconds=60)
-                ).first()
-
-                if recent:
-                    return JsonResponse({
-                        "success": False,
-                        "message": "You recently submitted this survey. Please wait before submitting again."
-                    })
-
-                survey_response = SurveyResponse.objects.create(
-                    student=student,
-                    survey_template=survey_template,
-                    flagged=False  # Will update this after checking responses
-                )
-            else:
-                recent = SurveyResponse.objects.select_for_update().filter(
-                    anonymous_student=ano_student,
-                    survey_template=survey_template,
-                    created__gte=timezone.now() - timezone.timedelta(seconds=60)
-                ).first()
-
-                if recent:
-                    return JsonResponse({
-                        "success": False,
-                        "message": "You recently submitted this survey. Please wait before submitting again."
-                    })
-
-                survey_response = SurveyResponse.objects.create(
-                    anonymous_student=ano_student,
-                    survey_template = survey_template,
-                    flagged=False
-                )
-            
-            # Create individual question responses
-            should_flag = False
-            for question in questions:
-                question_id = str(question.id)
-                response_value = request.data.get(question_id)
-                
-                if question.question_type == 'likert':
-                    likert_value = int(response_value)
-                    text_response = None
-                    # Check if this response should trigger flagging
-                    if likert_value >= 3:  # Assuming 3+ is concerning for any question
-                        should_flag = True
-                else:  # text response
-                    likert_value = None
-                    text_response = response_value
-                
-                QuestionResponse.objects.create(
-                    survey_response=survey_response,
-                    question=question,
-                    likert_value=likert_value,
-                    text_response=text_response
-                )
-            
-            # Update flagged status if needed
-            if should_flag:
-                survey_response.flagged = True
-                survey_response.save()
-            
-            # Pass question IDs instead of model objects
-            question_ids = [q.id for q in questions]
-            transaction.on_commit(
-                lambda: analyze_survey_responses_async.delay(survey_response.id, question_ids)
-            )
-        # Return success response
-        response_data = {
-            "success": True,
-            "message": "Thank you for your honest response! Your input makes a difference.",
-            "redirect_url": "/",
-            "data": SurveyResponseSerializer(survey_response).data,
-        }
-        return JsonResponse(response_data)
-    
-    except Exception as e:
-        # Return error response if any errors occur
-        logger.error("Error saving survey response: %s", str(e))
-        return JsonResponse({
-            "success": False,
-            "message": "There was an error with your submission.",
-        })
-
-@api_view(["GET"])
-def get_user_survey_questions(request, hash_link=None):
-    """
-    API endpoint to get survey questions for the current user based on their institution
-    @params:
-        request (HttpRequest): The HTTP request object
-        hash_link (str): Unique hash identifier for the survey template
-    
-    @returns:
-        JsonResponse: JSON response containing survey questions
-        - success (bool): True if questions retrieved successfully, False otherwise
-        - questions (list): List of question objects with details
-        - survey_template (dict): Survey template information
-        - error (str): Error message (on failure)
-    """
-    
-    # Handle hash link requests (authentication required)
-    if hash_link:
-        try:
-            survey_template = SurveyTemplate.objects.get(hash_link=hash_link)
-            
-            # Get all questions for this template
-            questions = SurveyQuestion.objects.filter(survey_template=survey_template).order_by('order')
-            
-            if not questions.exists():
-                return JsonResponse({
-                    "success": False, 
-                    "error": "No survey questions found for this template"
-                }, status=404)
-            
-            # Serialize the questions
-            questions_data = [{
-                'id': q.id,
-                'text': q.question_text,
-                'type': q.question_type,
-                'category': q.category,
-                'answer_choices': q.answer_choices,
-                'order': q.order
-            } for q in questions]
-            
-            return JsonResponse({
-                "success": True, 
-                "template_id": survey_template.id,
-                "questions": questions_data
-            })
-            
-        except SurveyTemplate.DoesNotExist:
-            return JsonResponse({
-                "success": False, 
-                "error": "Survey template not found"
-            }, status=404)
-
-    if not request.user.is_authenticated:
-        return JsonResponse({"success": False, "error": "Authentication required"})
-    try:
-        survey_template = None
-        
-        # If user is a student, get template by institution
-        if not request.user.is_superuser:
-            if request.user.institution_details:
-                survey_template = SurveyTemplate.objects.filter(
-                    institution=request.user.institution_details,
-                    used=True
-                ).first()
-                
-            if not survey_template:
-                survey_template = SurveyTemplate.objects.filter(
-                    institution=request.user.institution_details,
-                ).order_by('id').first()
-        
-        # If user is a superuser, get any template (for testing)
-        elif request.user.is_superuser:
-            survey_template = SurveyTemplate.objects.first()
-        
-        if not survey_template:
-            return JsonResponse({
-                "success": False, 
-                "error": "No survey template found for your institution"
-            }, status=404)
-        
-        # Get all questions for this template
-        questions = SurveyQuestion.objects.filter(survey_template=survey_template).order_by('order')
-        
-        # Check if questions exist
-        if not questions.exists():
-            return JsonResponse({
-                "success": False, 
-                "error": "No survey questions found for this template"
-            }, status=404)
-        
-        # Serialize the questions
-        questions_data = [{
-            'id': q.id,
-            'text': q.question_text,
-            'type': q.question_type,
-            'category': q.category,
-            'answer_choices': q.answer_choices,
-            'order': q.order
-        } for q in questions]
-        
-        # Final check - make sure we have questions data
-        if not questions_data:
-            return JsonResponse({
-                "success": False, 
-                "error": "Survey questions could not be loaded"
-            }, status=500)
-        
-        return JsonResponse({
-            "success": True, 
-            "template_id": survey_template.id,
-            "questions": questions_data
-        })
-        
-    except Exception as e:
-        # Log the error for debugging
-        import logging
-        logger = logging.getLogger(__name__)
-        logger.error(f"Error fetching survey questions: {str(e)}")
-        
-        return JsonResponse({
-            "success": False, 
-            "error": "An error occurred while loading survey questions"
-        }, status=500)
 
 @api_view(["GET"])
 def dashboard_api(request):
@@ -747,306 +282,6 @@ def dashboard_api(request):
 
     return JsonResponse(context)
 
-@api_view(["POST"])
-def login_view(request):
-    """
-    Authenticates users and establishes session with secure cookies.
-    
-    This endpoint handles user authentication by validating email/password credentials,
-    creating a session, and setting secure authentication cookies for frontend use.
-    
-    @params:
-        request (HttpRequest): Django HTTP request object
-            - method: Must be POST
-            - body (JSON): {
-                "email": str,     # User's email address
-                "password": str   # User's password
-            }
-    
-    @returns:
-        JsonResponse: Authentication result with user information and cookies
-        
-        Success Response (200):
-        {
-            "success": true,
-            "message": "Login successful!",
-            "is_admin": bool,              # Whether user is institution admin
-            "redirect_path": str           # Suggested redirect path based on user role
-        }
-        
-        Cookies Set on Success:
-        - auth_token: Session key (7 days, Lax SameSite)
-        - user_email: User's email (7 days, Lax SameSite)
-        - is_superuser: "true"/"false" (7 days, Lax SameSite)
-        - is_institution_admin: "true"/"false" (7 days, Lax SameSite)
-        
-        Error Responses:
-        - 400: {"error": "No data provided"} - Empty request body
-        - 400: {"error": "Email and password required"} - Missing credentials
-        - 400: {"error": "Invalid JSON format"} - Malformed JSON
-        - 400: {"success": false, "message": "Invalid credentials"} - Authentication failed
-        - 405: {"error": "Use POST method"} - Wrong HTTP method
-        - 500: {"error": "Server error occurred"} - Internal server error
-    
-    @notes:
-        - Cookie settings use Django settings for domain, security, and SameSite policies
-        - Redirect path is "/dashboard/" for admins, "/survey/" for students
-        - All cookies expire after 7 days
-        - Uses Django's built-in authentication system
-    """
-    if request.method == "POST":
-        try:
-            # Check if body exists
-            if not request.body:
-                return JsonResponse({'error': 'No data provided'}, status=400)
-            
-            # Parse JSON
-            data = json.loads(request.body)
-            email = data.get('email')
-            password = data.get('password')
-            
-            # Check if email/password provided
-            if not email or not password:
-                return JsonResponse({'error': 'Email and password required'}, status=400)
-            
-            user = authenticate(request, email=email, password=password)
-            
-            if user is not None:
-                login(request, user)
-                
-                response = JsonResponse({
-                    'success': True,
-                    'message': 'Login successful!',
-                    'is_admin': user.role == User.Role.INSTITUTION_ADMIN,
-                    'redirect_path': '/dashboard/' if user.role == User.Role.INSTITUTION_ADMIN else '/survey/'
-                })
-                
-                # Production-ready cookie settings
-                response.set_cookie(
-                    'auth_token', 
-                    request.session.session_key, 
-                    max_age=3600*24*7,
-                    path='/',
-                    domain=settings.COOKIE_DOMAIN,
-                    secure=settings.COOKIE_SECURE,
-                    httponly=False,
-                    samesite='Lax'
-                )
-                response.set_cookie(
-                    'user_email', 
-                    email, 
-                    max_age=3600*24*7,
-                    path='/',
-                    domain=settings.COOKIE_DOMAIN,
-                    secure=settings.COOKIE_SECURE,
-                    httponly=False,
-                    samesite='Lax'
-                )
-                response.set_cookie(
-                    'is_superuser', 
-                    'true' if user.is_superuser else 'false', 
-                    max_age=3600*24*7,
-                    path='/',
-                    domain=settings.COOKIE_DOMAIN,
-                    secure=settings.COOKIE_SECURE,
-                    httponly=False,
-                    samesite='Lax'
-                )
-                response.set_cookie(
-                    'is_institution_admin', 
-                    'true' if user.role == User.Role.INSTITUTION_ADMIN else 'false', 
-                    max_age=3600*24*7,
-                    path='/',
-                    domain=settings.COOKIE_DOMAIN,
-                    secure=settings.COOKIE_SECURE,
-                    httponly=False,
-                    samesite='Lax'
-                )
-                
-                return response
-            else:
-                return JsonResponse({'success': False, 'message': 'Invalid credentials'}, status=400)
-                
-        except json.JSONDecodeError:
-            return JsonResponse({'error': 'Invalid JSON format'}, status=400)
-        except Exception as e:
-            print(f"Login error: {e}")  # This will show in Django console
-            return JsonResponse({'error': 'Server error occurred'}, status=500)
-    return JsonResponse({'error': 'Use POST method'}, status=405)
-
-@api_view(["POST"])
-def register_view(request):
-    """
-    Registers new student users with institution validation.
-    
-    This endpoint creates new student accounts after validating institution membership
-    through email pattern matching and ensuring the institution exists in the system.
-    
-    @params:
-        request (HttpRequest): Django HTTP request object
-            - method: Must be POST
-            - data (JSON or form): {
-                "institution_name": str,    # Name of the institution
-                "name": str,               # Student's full name
-                "email": str,              # Student's email address
-                "password": str,           # Desired password
-                "confirm_password": str    # Password confirmation
-            }
-    
-    @returns:
-        Response: Registration result with success/error information
-        
-        Success Response (201):
-        {
-            "success": true,
-            "message": "Registration successful! Please log in."
-        }
-        
-        Error Responses:
-        - 400: {"success": false, "message": "Passwords do not match."}
-        - 400: {"success": false, "message": "Email already registered."}
-        - 400: {"success": false, "message": "Institution does not exist"}
-        - 400: {"success": false, "message": "Email does not match institution's format"}
-        - 500: {"success": false, "message": "Registration failed. Please try again."}
-    
-    @notes:
-        - Email must match the institution's regex pattern for validation
-        - Creates student users with role=STUDENT
-        - Institution must exist in the database before registration
-        - Uses case-insensitive regex matching for email validation
-        - Automatically associates user with the specified institution
-    """
-    if request.method == "POST":
-        # Extract form data from request.data (for JSON) or request.POST (for form data)
-        data = request.data if request.data else request.POST
-        
-        institution_name = data.get("institution_name")
-        name = data.get("name")
-        email = data.get("email")
-        password = data.get("password")
-        confirm_password = data.get("confirm_password")
-
-        # Validation checks
-        if password != confirm_password:
-            return Response({
-                'success': False,
-                'message': "Passwords do not match."
-            }, status=status.HTTP_400_BAD_REQUEST)
-
-        if User.objects.filter(email=email).exists():
-            return Response({
-                'success': False,
-                'message': "Email already registered."
-            }, status=status.HTTP_400_BAD_REQUEST)
-        
-        # Check if the institution exists
-        institution_details = Institution.objects.filter(institution_name=institution_name)
-        if not institution_details.exists():
-            return Response({
-                'success': False,
-                'message': "Institution does not exist"
-            }, status=status.HTTP_400_BAD_REQUEST)
-        
-        # Check if the email address matches the institution's pattern
-        institution_details = institution_details.first()
-        match_object = re.fullmatch(institution_details.institution_regex_pattern, email, re.IGNORECASE)
-        if not match_object:
-            return Response({
-                'success': False,
-                'message': "Email does not match institution's format"
-            }, status=status.HTTP_400_BAD_REQUEST)
-        
-        # Create a new student
-        try:
-            with transaction.atomic():
-                student = User.objects.create_student(
-                    email=email, 
-                    password=password, 
-                    institution_details=institution_details, 
-                    name=name
-                )
-                #student.save()
-                
-                return Response({
-                    'success': True,
-                    'message': "Registration successful! Please log in."
-                }, status=status.HTTP_201_CREATED)
-        except IntegrityError:
-            return Response({
-                'success': False,
-                'message': "Registration failed. Email already registered"
-            }, status=status.HTTP_400_BAD_REQUEST)
-            
-        except Exception as e:
-            return Response({
-                'success': False,
-                'message': "Registration failed. Please try again."
-            }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-@api_view(["POST"])
-def logout_view(request):
-    """
-    Logs out authenticated users and clears session cookies.
-    
-    This endpoint terminates the user's session and removes all authentication-related
-    cookies to ensure complete logout from the system.
-    
-    @params:
-        request (HttpRequest): Django HTTP request object
-            - method: Must be POST
-            - user: Must be authenticated (automatic Django middleware)
-    
-    @returns:
-        JsonResponse: Logout confirmation with cleared cookies
-        
-        Success Response (200):
-        {
-            "success": true,
-            "message": "You have been logged out successfully."
-        }
-        
-        Cookies Cleared:
-        - sessionid: Django session cookie
-        - auth_token: Custom authentication token
-        - user_email: Stored user email
-        - is_institution_admin: Admin status flag
-        - is_superuser: Superuser status flag
-    
-    @notes:
-        - Always returns success response regardless of authentication status
-        - Clears cookies using Django settings for domain configuration
-        - Uses Django's built-in logout() function to terminate session
-        - Safe to call multiple times or when already logged out
-    """
-    logout(request)
-
-    # Create response with success message
-    response_data = {
-        "success": True,
-        "message": "You have been logged out successfully."
-    }
-    
-    response = JsonResponse(response_data)
-    
-    # Clear auth cookies
-    cookies_to_clear = [
-        'sessionid',
-        'auth_token', 
-        'user_email', 
-        'is_institution_admin', 
-        'is_superuser'
-    ]
-    
-    for cookie_name in cookies_to_clear:
-        response.delete_cookie(
-            cookie_name, 
-            path="/", 
-            domain=settings.COOKIE_DOMAIN,
-        )
-    
-    return response
-
-
 # NOTE: API views are WIP. Need to add parameters for filtering db records
 @api_view(['GET'])
 def student_response_view(request):
@@ -1106,7 +341,6 @@ def student_response_view(request):
     else:
         return HttpResponseBadRequest("Request method not allowed")
 
-
 @api_view(["GET"])
 def flagged_responses_view(request):
     """
@@ -1164,7 +398,7 @@ def flagged_responses_view(request):
     
     else:
         return HttpResponseBadRequest("Request method not allowed")
-    
+
 @api_view(["GET"])
 def students_view(request):
     """
@@ -1649,7 +883,6 @@ def survey_questions_view(request, template_id):
         except Exception as e:
             return JsonResponse({"success": False, "error": str(e)})
 
-
 @api_view(["POST"])
 def use_template(request, template_id):
     """
@@ -1736,179 +969,3 @@ def get_active_template(institution):
                 template.save()
     
     return template
-
-@api_view(["POST"])
-def survey_autosave(request):
-    """
-    Saves survey progress to cache for authenticated users.
-    
-    This endpoint allows students to save their survey progress temporarily,
-    enabling them to resume later without losing their responses.
-    
-    @params:
-        request (HttpRequest): Django HTTP request object
-            - method: Must be POST
-            - user: Must be authenticated
-            - data (JSON): {
-                "template_id": int,        # Survey template ID
-                "answers": dict            # Question ID -> response mapping
-            }
-    
-    @returns:
-        JsonResponse: Save operation result
-        
-        Success Response (200):
-        {
-            "success": true,
-            "message": "Progress saved"
-        }
-        
-        Error Responses:
-        - 200: {"success": false, "message": "User not authorized"} - Not authenticated
-        - 200: {"success": false, "message": "Error: Wrong template id"} - Missing template_id
-        - 404: {"success": false, "message": "Invalid survey template"} - Template not found
-        - 200: {"success": false, "message": "Failed to save progress"} - Cache error
-    
-    @notes:
-        - Data cached for 30 minutes (1800 seconds)
-        - Cache key format: "survey_autosave_{email}_{template_id}"
-        - Includes timestamp of last save
-        - Overwrites previous autosave data for same user/template
-        - Uses Django cache framework for temporary storage
-    """
-    try:
-        if not request.user.is_authenticated:
-            return JsonResponse({"success": False, "message": "User not authorized"}, status=200)
-        data = request.data
-        template_id = request.data.get("template_id")  
-        if not template_id:
-            return JsonResponse({"success": False, "message": "Error: Wrong template id"}, status=200) 
-        try: 
-            SurveyTemplate.objects.get(id=template_id)
-        except SurveyTemplate.DoesNotExist:
-            return JsonResponse({
-                "success": False, 
-                "message": "Invalid survey template"
-            }, status=404)     
-        answers = request.data.get("answers", {})
-        student_name, school_email = request.user.name, request.user.email
-        cache_data = {
-            "template_id": template_id, 
-            "student_name": student_name if student_name else "", 
-            "school_email": school_email, 
-            "last_saved": datetime.now().isoformat(),
-            "answers": answers, 
-        }
-        cache.set(f"survey_autosave_{school_email}_{template_id}", json.dumps(cache_data), timeout=1800)
-        return JsonResponse({"success": True, "message": "Progress saved"})
-    except Exception as e:
-        logger.error(f"Autosave error: {str(e)}")
-        return JsonResponse({"success": False, "message": "Failed to save progress"}, status=200)
-
-
-@api_view(["GET"])
-def survey_autosave_load(request, template_id):
-    """
-    Loads previously saved survey progress from cache.
-    
-    This endpoint retrieves autosaved survey data for authenticated users,
-    allowing them to resume their survey from where they left off.
-    
-    @params:
-        request (HttpRequest): Django HTTP request object
-            - method: Must be GET
-            - user: Must be authenticated
-        template_id (int): Survey template ID to load progress for
-    
-    @returns:
-        JsonResponse: Load operation result with saved data
-        
-        Success Response (200):
-        {
-            "success": true,
-            "saved_data": {
-                "template_id": int,
-                "student_name": str,
-                "school_email": str,
-                "last_saved": str,         # ISO datetime string
-                "answers": dict            # Question ID -> response mapping
-            }
-        }
-        
-        Error Responses:
-        - 200: {"success": false, "message": "User not authorized"} - Not authenticated
-        - 200: {"success": false, "message": "No autosaved data found"} - No cached data
-        - 200: {"success": false, "message": "Corrupted save data. Please press the clear button"} - Invalid JSON
-        - 200: {"success": false, "message": "Failed to load autosave"} - Cache error
-    
-    @notes:
-        - Cache key format: "survey_autosave_{email}_{template_id}"
-        - Automatically clears corrupted cache data
-        - Returns user's name and email along with answers
-        - Data expires after 30 minutes of inactivity
-    """
-    try:
-        # Check authentication
-        if not request.user.is_authenticated:
-            return JsonResponse({"success": False, "message": "User not authorized"}, status=200)
-        school_email = request.user.email
-        cache_key = f"survey_autosave_{school_email}_{template_id}"
-        cached_data = cache.get(cache_key)
-        if cached_data:
-            try:
-                return JsonResponse({"success": True, "saved_data": json.loads(cached_data)}, status=200)
-            except json.JSONDecodeError:
-                # Corrupted data
-                cache.delete(cache_key)
-                return JsonResponse({"success": False, "message": "Corrupted save data. Please press the clear button"}, status=200)
-        else:
-            return JsonResponse({"success": False, "message": "No autosaved data found"}, status=200)
-    except Exception as e:
-        logger.error(f"Autosave load error: {str(e)}")
-        return JsonResponse({"success": False, "message": "Failed to load autosave"}, status=200)
-
-@api_view(["DELETE"])
-def survey_autosave_clear(request, template_id):
-    """
-    Clears saved survey progress from cache.
-    
-    This endpoint allows authenticated users to manually delete their autosaved
-    survey progress, useful when starting fresh or after successful submission.
-    
-    @params:
-        request (HttpRequest): Django HTTP request object
-            - method: Must be DELETE
-            - user: Must be authenticated
-        template_id (int): Survey template ID to clear progress for
-    
-    @returns:
-        JsonResponse: Clear operation result
-        
-        Success Response (200):
-        {
-            "success": true,
-            "message": "Autosave data cleared"
-        }
-        
-        Error Responses:
-        - 200: {"success": false, "message": "User not authorized"} - Not authenticated
-        - 200: {"success": false, "message": "Failed to clear draft"} - Cache error
-    
-    @notes:
-        - Cache key format: "survey_autosave_{email}_{template_id}"
-        - Safe to call even if no autosave data exists
-        - Typically called after successful survey submission
-        - Immediately removes data from cache
-    """
-    try:
-        if not request.user.is_authenticated:
-            return JsonResponse({"success": False, "message": "User not authorized"}, status=200)
-        school_email = request.user.email
-        cache_key = f"survey_autosave_{school_email}_{template_id}"
-        
-        cache.delete(cache_key)
-
-        return JsonResponse({"success": True, "message": "Autosave data cleared"})
-    except Exception as e:
-        logger.error(f"Clear autosave error: {str(e)}")
-        return JsonResponse({"success": False, "message": "Failed to clear draft"}, status=200)
